excel-automation/requirements.txt
pandas==2.1.4
openpyxl==3.1.2
python-dotenv==1.0.0
schedule==1.2.0
sqlalchemy==2.0.25
pyodbc==5.0.1
requests==2.31.0
